Fix various serialization issues

+++ include/swift/SIL/SILInstruction.h
@@ -7648,14 +7648,14 @@ class AutoDiffFunctionExtractInst :
                            SingleValueInstruction> {
 public:
   struct Extractee {
-    enum innerty : uint8_t {
+    enum innerty : unsigned {
       Original = 0,
       JVP = 1,
       VJP = 2
     } rawValue;
     Extractee() = default;
     Extractee(innerty rawValue) : rawValue(rawValue) {}
-    Extractee(uint8_t rawValue) : Extractee((innerty)rawValue) {}
+    Extractee(unsigned rawValue) : Extractee((innerty)rawValue) {}
     explicit Extractee(StringRef name);
     operator innerty() const { return rawValue; }
   };
